fix: image conversion for other image modes (skip alpha band) (#712)
pillow image modes: https://pillow.readthedocs.io/en/stable/handbook/concepts.html

In my case, i was dealing with a `P` mode image.  and got hit with `IndexError: tuple index out of range`. While the current operation is useful it's not needed for cases such as the one I faced. So we put it under a condition.